Add baseline grid and standardize code comment styles (#1)
* Creates first draft of baseline grid

* fix the .u-textLarge line-height issue

* Fixes line height and sizing issue with .u-textLarge

Increases the fontsize of large text from 18px to 21px with a line height of 3rems. This helps the large text sit better on a readable leading.

* Makes baseline grid and font more generic for wider use.

The previous iteration was specific to an implementation, and this makes it slightly more accessible for wider use

* Uses system defaults for all typography

- Uses system default for the type family
- Uses pseudo classes for borders to keep the baseline grid from being modified when using borders

* Fixes improper border declaration

* Adds robust border utils for each box side, plus a reset

* reverts comment annotation from process

* Adds default button element to component

To be consistent with rules and other components that can also interact with their respective semantic elements. The class should always still be used, but in cases where other libraries are introduced, this will override those defaults.

* Standardizes code comments

- Creates a standard code comment style
- Adds placeholder files for leader and lists
- Separates clearfix and float from the layout file

* removes leader, moves list to utils

* use padding instead of margin